Filters

+++ backend/api/views.py
@@ -1,9 +1,13 @@
 from django.shortcuts import get_object_or_404
 from django_filters.rest_framework import DjangoFilterBackend
-from rest_framework import viewsets, permissions, decorators, status
+from rest_framework import viewsets, permissions, decorators, status, filters
 from rest_framework.response import Response
 from djoser.views import UserViewSet
 
+from .filters import (
+    IngredientFilter,
+    RecipeFilter
+)
 from .permissions import (
     IsAuthorOrReadOnly,
     IsAdminOrReadOnly
@@ -59,17 +63,30 @@ class CustomUserViewSet(UserViewSet):
         author = get_object_or_404(CustomUser, id=id)
 
         if request.method == 'POST':
+
+            if user.id == author.id:
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+
+            if Subscribe.objects.filter(user=user, author=author).exists():
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+
             Subscribe.objects.create(user=user, author=author)
             serializer = SubscribeSerializer(
                 author, context={'request': request}
             )
+
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         
         if request.method == 'DELETE':
+
+            if not Subscribe.objects.filter(user=user, author=author).exists():
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+
             subsctiption = get_object_or_404(
                 Subscribe, user=user, author=author
             )
             subsctiption.delete()
+
             return Response(status=status.HTTP_204_NO_CONTENT)
 
 
@@ -84,14 +101,17 @@ class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Ingredient.objects.all()
     serializer_class = IngredientSerializer
     permission_classes = (permissions.AllowAny,)
+    filter_backends = (DjangoFilterBackend, filters.SearchFilter,)
+    filterset_class = IngredientFilter
+    search_fields = ('name',)
     pagination_class = None
-    # filter_backends = 
 
 
 class RecipeViewSet(viewsets.ModelViewSet):
     queryset = Recipe.objects.all()
     permission_classes = (IsAuthorOrReadOnly, IsAdminOrReadOnly,)
-    # filter_backends = 
+    filter_backends = (DjangoFilterBackend,)
+    filterset_class = RecipeFilter
 
     def get_serializer_class(self):
         if self.request.method == 'GET':
@@ -114,15 +134,24 @@ class RecipeViewSet(viewsets.ModelViewSet):
         recipe = get_object_or_404(Recipe, pk=pk)
 
         if self.request.method == 'POST':
+
+            if Favorite.objects.filter(user=user, recipe=recipe).exists():
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+
             Favorite.objects.create(user=user, recipe=recipe)
             serializer = RecipeShowSerializer(
                 recipe, context={'request': request}
             )
+
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         
         if self.request.method == 'DELETE':
+            if not Favorite.objects.filter(user=user, recipe=recipe).exists():
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+
             favorite = Favorite.objects.get(user=user, recipe=recipe)
             favorite.delete()
+
             return Response(status=status.HTTP_204_NO_CONTENT)
 
     @decorators.action(
@@ -135,15 +164,25 @@ class RecipeViewSet(viewsets.ModelViewSet):
         recipe = get_object_or_404(Recipe, pk=pk)
 
         if self.request.method == 'POST':
+
+            if ShoppingCart.objects.filter(user=user, recipe=recipe).exists():
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+
             ShoppingCart.objects.create(user=user, recipe=recipe)
             serializer = RecipeShowSerializer(
                 recipe, context={'request': request}
             )
+
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         
         if self.request.method == 'DELETE':
+
+            if not ShoppingCart.objects.filter(user=user, recipe=recipe).exists():
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+
             shopping_cart = ShoppingCart.objects.get(user=user, recipe=recipe)
             shopping_cart.delete()
+
             return Response(status=status.HTTP_204_NO_CONTENT)
     
     @decorators.action(
